requirements.txt
numpy
pandas
matplotlib.pyplot
seaborn
sklearn
xgboost
shap
tensorflow
